requirements.txt
torch 
torchaudio 
transformers 
numpy 
onnxruntime 
funasr
llama-cpp-python
pydub